pyproject.toml
[tool.poetry]
name = "ai-music-composer"
version = "0.1.0"
description = "An intelligent music composition system with MIDI output"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "composer", from = "src"}]

[tool.poetry.scripts]
composer = "composer_cli:main"

[tool.poetry.dependencies]
python = "^3.9"
mido = "^1.3.0"
click = "^8.1.0"
rich = "^13.0.0"
langchain = "^0.1.0"
langchain-openai = "^0.0.5"
python-dotenv = "^1.0.0"
pydantic = "^2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^6.2.5"
black = "^24.1.1"
flake8 = "^7.0.0"
flake8-docstrings = "^1.7.0"
flake8-bugbear = "^24.1.17"
pre-commit = "^3.6.0"
isort = "^5.13.0"
autoflake = "^2.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
^/tests/
'''

[tool.flake8]
max-line-length = 88
extend-ignore = [
    "E203",  # Whitespace before ':' (conflicts with Black)
    "W503",  # Line break before binary operator (conflicts with Black)
    "E501",  # Line too long (handled by Black/autopep8)
]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    "*.egg-info",
]
max-complexity = 10
docstring-convention = "google"
per-file-ignores = [
    "__init__.py:D104",  # Missing docstring in public package
    "tests/*:D100,D101,D102,D103",  # Missing docstrings in tests
]

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["composer", "midi_generator", "llm_composer", "database"]

[tool.autoflake]
remove-all-unused-imports = true
remove-unused-variables = true
remove-duplicate-keys = true
ignore-init-module-imports = true
